Fix integer overflow in PagingPredicate with large pages
Fixes: https://github.com/hazelcast/hazelcast/issues/12079

+++ hazelcast/src/main/java/com/hazelcast/util/SortingUtil.java
@@ -147,9 +147,10 @@ public final class SortingUtil {
         int nearestPage = nearestAnchorEntry.getKey();
         int pageSize = pagingPredicate.getPageSize();
         int page = pagingPredicate.getPage();
-        int totalSize = pageSize * (page - nearestPage);
+        long totalSize = pageSize * ((long) page - nearestPage);
         if (list.size() > totalSize) {
-            list = list.subList(0, totalSize);
+            // it's safe to cast totalSize back to int here since it's limited by the list size
+            list = list.subList(0, (int) totalSize);
         }
         return list;
     }
@@ -167,17 +168,18 @@ public final class SortingUtil {
         int nearestPage = nearestAnchorEntry.getKey();
         int page = pagingPredicate.getPage();
         int pageSize = pagingPredicate.getPageSize();
-        int begin = pageSize * (page - nearestPage - 1);
+        long begin = pageSize * ((long) page - nearestPage - 1);
         int size = list.size();
         if (begin > size) {
             return new ResultSet();
         }
-        int end = begin + pageSize;
+        long end = begin + pageSize;
         if (end > size) {
             end = size;
         }
         setAnchor(list, pagingPredicate, nearestPage);
-        List<Map.Entry> subList = list.subList(begin, end);
+        // it's safe to cast begin and end back to int here since they are limited by the list size
+        List<Map.Entry> subList = list.subList((int) begin, (int) end);
         return new ResultSet(subList, iterationType);
     }
